refactor: extract web_server Pydantic models to web_models.py (pure schema move)
